Clarify the GNU General Public License version.
Section 14 of the GNU General Public License (see LICENSE.txt) specifies that
the source code must specify the version of the license it's under or the person
using the source code can pick any version of the GPL, weakening the copyleft.

The source code may also specify that any later version can be used to allow
people to use the code under a newer version of the GPL, but also the current
version. This is particularly useful for switching to a newer version, or
letting others use this code in works licensed under a newer version.

This commit clarifies unambiguously the version of the license. I didn't bother
clarifying the outputted .js file as it's object code under the terms of the
license and not source code.